LogisTrack/backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
djangorestframework-simplejwt==5.3.1
openpyxl==3.1.5
